VSCode Transfer Complete
Arduino's IDE was getting crouded and annoying, so moving to VSCode for a better IDE and to help with easier Multi-file coding. It also makes C++ a bit more true to form instead of the Arduino flavor. Because of this, some code had to be changed around so that it works nicer in VSCode. Also the file structure changed as it uses Platform IO.