Production PCB Added
Production PCB Added it was noticed since prodution there is an issue with the PCB that needs to be addressed in a future update. For now it can be fixed mostly by ripping the resistor off the board. Ideally then a wire will be soldered in its place to 3.3V