PCB V2 & Case Work & Code Updates
PCB Version 2 started and completed. This makes it smaller and should fix the issues with having to solder a patch wire to the board. It also makes pin 13 no longer used for controlling the connected RGB components as it should not glow constantly which could disrupt RGB patterns. The new PCB also uses a much better mounting system for the Molex power connector. It also increased the trace width for the power traces to increase the current capabilities. As well a new case was started for covering the PCB inside a computer. Lastly, the Default pattern was bug fixed to no longer have issues with the MAXPIXELS constant being under 255. It also reflects the PCB change for pin 13.