New Patterns and New Features and Deployment
Some changes as it was fully deployed and I was able to test everything. I was concerned about the temperature of the teensy so I added another serial command to check the internal temperature of the teensy. I was also testing out a ton of patterns major trial and error.